moving shardinginterface implementation from tensor to mesh lib
